extract useNavPrefetch hook
